Start of rendering logic
Differentiated somewhat between Minecraft and Hero code, but they are so tightly tied it will be tricky to have logical code organization that doesn't jump between files too much. Will come back to this later and clean it up.

+++ src/main/java/gr8pefish/heroreactions/hero/data/HeroData.java
@@ -12,7 +12,7 @@ public class HeroData {
     //Reaction -> count
     private static final ConcurrentHashMap<FeedbackTypes, Integer> feedbackActivity = new ConcurrentHashMap<>();
     //Reaction -> ratio (count/total)
-    private static final ConcurrentHashMap<FeedbackTypes, Float> feedbackRatios = new ConcurrentHashMap<>();
+    private static final ConcurrentHashMap<FeedbackTypes, Double> feedbackRatios = new ConcurrentHashMap<>();
 
     public static class Viewers {
         public static int direct;
@@ -35,7 +35,7 @@ public class HeroData {
             return feedbackActivity;
         }
         //Use feedbackRatios map
-        public static ConcurrentHashMap<FeedbackTypes, Float> getFeedbackRatios() {
+        public static ConcurrentHashMap<FeedbackTypes, Double> getFeedbackRatios() {
             return feedbackRatios;
         }
